Add files via upload
Machine learning (Multiple regression) using sklearn and some data visualisation. This regression model looks to predict the % chance of success that a student may have in obtaining admission to the IEEE Master's program based on a number of factors. For anyone interest on the full data set it can be found on Kaggle via this link; https://www.kaggle.com/mohansacharya/graduate-admissions